feat: clean code
