+++ internal/index.ts
@@ -1,10 +1,8 @@
 import { css } from './plugins/css'
 import { analyze } from './plugins/analyze'
 
-// , analyze
 export const internalPlugins = { css, analyze }
 
-export * from './bundle'
 export * from './process'
 export * from './helpers/gen'
 export * from './helpers/format'
