+++ packages/vue/package.json
@@ -49,7 +49,8 @@
     "name": "fect",
     "formats": [
       "esm",
-      "cjs"
+      "cjs",
+      "iife"
     ]
   }
 }
